Added password change option and email validation in user settings. User management self delete now logs user out. More snackbars added including admin to admin delete error.

+++ lib/screens/user_settings.dart
@@ -25,12 +25,14 @@ class UserSettingsScreenState extends State<UserSettingsScreen> {
   final _formKey = GlobalKey<FormState>();
   final nameController = TextEditingController();
   final emailController = TextEditingController();
+  final passwordController = TextEditingController();
 
   @override
   void dispose() {
     // Clean up the controller when the widget is disposed.
     nameController.dispose();
     emailController.dispose();
+    passwordController.dispose();
     super.dispose();
   }
 
@@ -93,8 +95,16 @@ class UserSettingsScreenState extends State<UserSettingsScreen> {
                   //initialValue: "Password",
                   // The validator receives the text that the user has entered.
                   validator: (value) {
+                    // if (value == null || value.isEmpty) {
+                    //   return 'Please enter some text';
+                    // }
                     if (value == null || value.isEmpty) {
-                      return 'Please enter some text';
+                      return null;
+                    }
+                    if (!RegExp(
+                            r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
+                        .hasMatch(value)) {
+                      return "Invalid Email";
                     }
                     return null;
                   },
@@ -104,43 +114,71 @@ class UserSettingsScreenState extends State<UserSettingsScreen> {
                 width: double.infinity,
                 height: MediaQuery.of(context).size.height / 16,
               ),
+              Container(
+                width: double.infinity,
+                child: TextFormField(
+                  // initialValue: routeArguments['blog_body'],
+                  minLines: 1,
+                  maxLines: 1,
+                  //keyboardType: TextInputType.,
+                  controller: passwordController,
+                  obscureText: true,
+                  decoration: InputDecoration(
+                      hintText: 'New Password', border: OutlineInputBorder()),
+
+                  //initialValue: "Password",
+                  // The validator receives the text that the user has entered.
+                  validator: (value) {
+                    // if (value == null || value.isEmpty) {
+                    //   return 'Please enter some text';
+                    // }
+                    return null;
+                  },
+                ),
+              ),
               Padding(
                 padding: const EdgeInsets.symmetric(vertical: 16.0),
                 child: ElevatedButton(
                   onPressed: () async {
-                    String resBody = await updateUserRequest(
-                      nameController.text,
-                      emailController.text,
-                    );
-                    //print(resBody);
-                    if (resBody != "error") {
-                      //postLoginSetup(resBody);
-                      //Navigator.pop(context);
-                      window.localStorage['name'] =
-                          nameController.text.isNotEmpty
-                              ? nameController.text
-                              : window.localStorage['name']!;
-
-                      window.localStorage['email'] =
-                          emailController.text.isNotEmpty
-                              ? emailController.text
-                              : window.localStorage['email']!;
-                      Navigator.pushNamed(context, '/profile');
-                    } else {
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        const SnackBar(
-                            content: Text('Account update unsuccessful')),
-                      );
-                    }
-
-                    // Validate returns true if the form is valid, or false otherwise.
                     if (_formKey.currentState!.validate()) {
-                      // If the form is valid, display a snackbar. In the real world,
-                      // you'd often call a server or save the information in a database.
-                      // ScaffoldMessenger.of(context).showSnackBar(
-                      //   const SnackBar(content: Text('Processing Data')),
-                      // );
+                      String resBody = await updateUserRequest(
+                        nameController.text,
+                        emailController.text,
+                        passwordController.text,
+                      );
+                      //print(resBody);
+                      if (resBody != "error") {
+                        //postLoginSetup(resBody);
+                        //Navigator.pop(context);
+                        window.localStorage['name'] =
+                            nameController.text.isNotEmpty
+                                ? nameController.text
+                                : window.localStorage['name']!;
 
+                        window.localStorage['email'] =
+                            emailController.text.isNotEmpty
+                                ? emailController.text
+                                : window.localStorage['email']!;
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          const SnackBar(content: Text('Account Updated!')),
+                        );
+                        Navigator.pushNamed(context, '/profile');
+                      } else {
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          const SnackBar(
+                              content: Text('Account update unsuccessful')),
+                        );
+                      }
+
+                      // Validate returns true if the form is valid, or false otherwise.
+                      if (_formKey.currentState!.validate()) {
+                        // If the form is valid, display a snackbar. In the real world,
+                        // you'd often call a server or save the information in a database.
+                        // ScaffoldMessenger.of(context).showSnackBar(
+                        //   const SnackBar(content: Text('Processing Data')),
+                        // );
+
+                      }
                     }
                   },
                   child: const Text('Update'),
@@ -171,17 +209,20 @@ class UserSettingsScreenState extends State<UserSettingsScreen> {
                               child: const Text('Delete'),
                               onPressed: () async {
                                 var res = await deleteUser(
-                                    window.localStorage['user_id']);
+                                    window.localStorage['user_id'], context);
                                 print(res);
+                                print("ok");
                                 // generateUsersList(callbackUpdateUsersList,
                                 //     usersPerPage: this.usersPerPage,
                                 //     page: this.page);
-                                window.localStorage.clear();
-                                //Navigator.of(context).pop();
-                                ScaffoldMessenger.of(context).showSnackBar(
-                                  const SnackBar(
-                                      content: Text('Account Deleted')),
-                                );
+                                if (res != "error") {
+                                  window.localStorage.clear();
+                                  //Navigator.of(context).pop();
+                                  ScaffoldMessenger.of(context).showSnackBar(
+                                    const SnackBar(
+                                        content: Text('Account Deleted')),
+                                  );
+                                }
                                 Navigator.pushNamed(context, '/login');
                               },
                             ),
